backend/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
pandas
python-multipart
pymupdf
rich
chromadb
pydantic-settings
python-dotenv
Pillow